Correct pallet schema types and test data
Schema clarification: PackagingData.packID NVARCHAR(2) is the record
identifier (no BIGINT identity needed). PalletPackages.packagingID
stores that same NVARCHAR(2) code. PalletMain.palletID and
PalletPackages.palletItemID need INT IDENTITY.

sql/migrate_logistics_pallets.sql (new)
- Drops PalletPackages, DeliveryLink, PalletMain in FK order
- Recreates PalletMain with palletID INT IDENTITY(1,1)
- Recreates DeliveryLink with INT palletID + FK + indexes
- Recreates PalletPackages with palletItemID INT IDENTITY(1,1)
  and packagingID NVARCHAR(2) (FK to PackagingData.packID)
- Note: migrate_logistics_packagingid.sql should NOT be run

sql/test_data_logistics.sql (rewritten)
- PackagingData: exact dataset from spec (SD/MD/LD/XL/XX drums,
  SB/MB/LB/XB/CB boxes, C1/C2/C3 custom cardboard)
- PalletData: PH/P8/PE/PS/PC/ST/GE with correct dimensions
- PalletValidation: exact allowed-packaging rules per pallet type
- Destinations + DeliveryMain: unchanged (5 customers, 7 open
  picksheets covering all date buckets)
- Removed example PalletMain/PalletPackages rows (use builder instead)

Routes
- palletmain.js: sql.BigInt → sql.Int for palletId throughout
- palletpackages.js: complete rewrite — sql.Int for palletId/
  palletItemId; packagingID as sql.NVarChar(2); GET /pallet/:id
  joins PackagingData on pd.packID = pp.packagingID
- palletvalidation.js GET /pallet/:id: joins PackagingData and
  returns packagingID (the code) + full PackagingData detail
- deliverymain.js POST /:id/pallets: palletId sql.BigInt → sql.Int

warehouse.js
- Radio button value = packagingID code (NVARCHAR) directly
- addPackage sends packType string; state stores packagingID code
- Reverts all BIGINT-related data-packid/data-desc workarounds

+++ routes/deliverymain.js
@@ -212,7 +212,7 @@ router.post('/:deliveryId/pallets', async (req, res) => {
         const pool = await getPool();
         await pool.request()
             .input('deliveryId', sql.BigInt, req.params.deliveryId)
-            .input('palletId',   sql.BigInt, palletId)
+            .input('palletId',   sql.Int,    palletId)
             .query(`INSERT INTO Logistics.dbo.DeliveryLink (deliveryID, palletID)
                     VALUES (@deliveryId, @palletId)`);
         res.status(201).json({ success: true });
